+++ server/src-lib/Hasura/Backends/MSSQL/Schema/IfMatched.hs
@@ -9,8 +9,10 @@ module Hasura.Backends.MSSQL.Schema.IfMatched
   )
 where
 
+import Data.Has
 import Data.Text.Extended
 import Hasura.Backends.MSSQL.Types.Insert
+import Hasura.Backends.MSSQL.Types.Internal (ScalarType (..))
 import Hasura.GraphQL.Parser
   ( InputFieldsParser,
     Kind (..),
@@ -18,6 +20,7 @@ import Hasura.GraphQL.Parser
     UnpreparedValue (..),
   )
 import Hasura.GraphQL.Parser qualified as P
+import Hasura.GraphQL.Parser.Class
 import Hasura.GraphQL.Schema.Backend
 import Hasura.GraphQL.Schema.BoolExp
 import Hasura.GraphQL.Schema.Common
@@ -64,7 +67,7 @@ ifMatchedObjectParser sourceName tableInfo maybeSelectPerms maybeUpdatePerms = r
   -- Short-circuit if we don't have sufficient permissions.
   selectPerms <- hoistMaybe maybeSelectPerms
   updatePerms <- hoistMaybe maybeUpdatePerms
-  matchColumnsEnum <- MaybeT $ tableSelectColumnsEnum sourceName tableInfo selectPerms
+  matchColumnsEnum <- MaybeT $ tableInsertMatchColumnsEnum sourceName tableInfo selectPerms
   updateColumnsEnum <- MaybeT $ tableUpdateColumnsEnum tableInfo updatePerms
 
   -- The style of the above code gives me some cognitive dissonance: We could
@@ -93,3 +96,46 @@ ifMatchedObjectParser sourceName tableInfo maybeSelectPerms maybeUpdatePerms = r
       _imUpdateColumns <-
         P.fieldWithDefault updateColumnsName Nothing (G.VList []) (P.list updateColumnsEnum)
       pure $ IfMatched {..}
+
+-- | Table insert_match_columns enum
+--
+-- Parser for an enum type that matches the columns that can be used
+-- for insert match_columns for a given table.
+-- Maps to the insert_match_columns object.
+--
+-- Return Nothing if there's no column the current user has "select"
+-- permissions for.
+tableInsertMatchColumnsEnum ::
+  forall m n r.
+  (MonadSchema n m, MonadRole r m, MonadTableInfo r m, Has P.MkTypename r) =>
+  SourceName ->
+  TableInfo 'MSSQL ->
+  SelPermInfo 'MSSQL ->
+  m (Maybe (Parser 'Both n (Column 'MSSQL)))
+tableInsertMatchColumnsEnum sourceName tableInfo selectPermissions = do
+  tableGQLName <- getTableGQLName @'MSSQL tableInfo
+  columns <- tableSelectColumns sourceName tableInfo selectPermissions
+  enumName <- P.mkTypename $ tableGQLName <> $$(G.litName "_insert_match_column")
+  let description =
+        Just $
+          G.Description $
+            "select match_columns of table " <>> tableInfoName tableInfo
+  pure $
+    P.enum enumName description
+      <$> nonEmpty
+        [ ( define $ pgiName column,
+            pgiColumn column
+          )
+          | column <- columns,
+            isMatchColumnValid column
+        ]
+  where
+    define name =
+      P.Definition name (Just $ G.Description "column name") P.EnumValueInfo
+
+-- | Check whether a column can be used for match_columns.
+isMatchColumnValid :: ColumnInfo 'MSSQL -> Bool
+isMatchColumnValid = \case
+  -- Unfortunately MSSQL does not support comparison for TEXT types.
+  ColumnInfo {pgiType = ColumnScalar TextType} -> False
+  _ -> True