+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 (Add highlights/major features below)
 
 - server: add request and response bodies to OpenAPI specification of REST endpoints
+- server: implement upsert mutations for MS SQL Server (close #7864)
 
 ### Bug fixes and improvements
 (Add entries below in the order of server, console, cli, docs, others)